New UMLs, changed Card template and tests
